Use authentication middleware
* remove userId from input and retrieve it from context
* change repository accordingly

+++ server/pkg/jwt/jwt.go
@@ -6,17 +6,13 @@ import (
 	"time"
 )
 
-// secret key being used to sign tokens
 var (
 	SecretKey = []byte("secret")
 )
 
-// GenerateToken generates a jwt token and assign a username to it's claims and return it
 func GenerateToken(userId int) (string, error) {
 	token := jwt.New(jwt.SigningMethodHS256)
-	/* Create a map to store our claims */
 	claims := token.Claims.(jwt.MapClaims)
-	/* Set token claims */
 	claims["user_id"] = userId
 	claims["exp"] = time.Now().Add(time.Hour * 24).Unix()
 	tokenString, err := token.SignedString(SecretKey)
@@ -27,13 +23,12 @@ func GenerateToken(userId int) (string, error) {
 	return tokenString, nil
 }
 
-// ParseToken parses a jwt token and returns the username in it's claims
 func ParseToken(tokenStr string) (int, error) {
 	token, err := jwt.Parse(tokenStr, func(token *jwt.Token) (interface{}, error) {
 		return SecretKey, nil
 	})
 	if claims, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
-		username := claims["user_id"].(int)
+		username := int(claims["user_id"].(float64))
 		return username, nil
 	} else {
 		return -1, err
